Naming convention change dashes with underscores. Added Copyrights
