+++ deployment/dev.env
@@ -23,11 +23,11 @@
 #
 #############################################################################
 
-DJANGO_SETTINGS_MODULE='statuto-territorio-rt.settings'
+DJANGO_SETTINGS_MODULE='statuto_territorio_rt.settings'
 DJANGO_ALLOWED_HOSTS='127.0.0.1 localhost' 
 DJANGO_DEBUG='True'
-DJANGO_SECRET_KEY='&z+bl66a=!^9#njhsxwcq8n#)hc2n_isw%x0w5ir%3)gk7w8!j'
-DJANGO_DATABASE_URL='postgres://geosolutions:geosolutions@localhost/statutoterritoriort'
+DJANGO_SECRET_KEY='abcdefgh'
+DJANGO_DATABASE_URL='postgres://USER:PASSWORD@HOST:PORT/NAME'
 DJANGO_PUBLIC_URL='http://127.0.0.1:8000'
 DJANGO_EMAIL_HOST='smtp.mail.com'
 DJANGO_EMAIL_HOST_USER='phony@mail.com'
