+++ requirements/requirements.txt
@@ -1,5 +1,6 @@
 dj-database-url==0.5.0
-Django==2.1.2
+Django==2.0.8
+django-filter==2.0.0
 psycopg2==2.7.5
 psycopg2-binary==2.7.5
 pytz==2018.5
