--- statuto-territorio-rt/base/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-
-def homeView(request):
-    return render(request, "base/home.html")
