+++ statuto_territorio_rt/base/urls.py
@@ -1,4 +1,14 @@
-"""statuto-territorio-rt URL Configuration
+#########################################################################
+#
+# Copyright 2018, GeoSolutions Sas.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+#
+#########################################################################
+
+"""statuto_territorio_rt URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.1/topics/http/urls/
@@ -13,10 +23,9 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
-from django.urls import path, include
+from django.urls import path
+from .views import homeView
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include('base.urls')),
+    path('', homeView, name='home_view'),
 ]
