+++ statuto_territorio_rt/statuto_territorio_rt/settings.py
@@ -1,5 +1,15 @@
+#########################################################################
+#
+# Copyright 2018, GeoSolutions Sas.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+#
+#########################################################################
+
 """
-Django settings for statuto-territorio-rt project.
+Django settings for statuto_territorio_rt project.
 
 Generated by 'django-admin startproject' using Django 2.1.2.
 
@@ -56,7 +66,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'statuto-territorio-rt.urls'
+ROOT_URLCONF = 'statuto_territorio_rt.urls'
 
 TEMPLATES = [
     {
@@ -74,7 +84,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'statuto-territorio-rt.wsgi.application'
+WSGI_APPLICATION = 'statuto_territorio_rt.wsgi.application'
 
 
 # Database
