+++ statuto_territorio_rt/statuto_territorio_rt/utils.py
